refactor out vs stuff
